requirements.txt
requests>=2.28.0
python-dateutil>=2.8.2
pytest>=7.0.0
python-dotenv>=1.0.0
aiogram==2.25.1
aiohttp==3.8.6
yookassa>=3.7.0